feature(navUI): add 'index' component with initial refactoring #550
